webapp/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
boto3==1.28.17
botocore==1.31.17
certifi==2023.7.22
cffi==1.15.1
circuitbreaker==1.4.0
cryptography==41.0.2
Django==4.2.4
django-s3-storage==0.14.0
djangorestframework==3.14.0
django-storages==1.13.2
jmespath==1.0.1
mysql-connector-python==8.1.0
oci==2.109.0
protobuf==4.21.12
pycparser==2.21
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==1.26.16